Move helper functions to a concern,
